Create seperate filter hook, create cookies tab component for pre-processing query
